refactor(capture): split repository attribution
